requirements.txt
fastapi
uvicorn
sqlmodel
passlib[argon2]
python-jose[cryptography]
python-multipart